feat: 예외가 발생했을 때 비정상 종료 되지 않고 재시작하도록 함
